+++ src/ts/Views/MRAID.ts
@@ -186,6 +186,17 @@ export class MRAID extends View {
         super.hide();
     }
 
+    public createMRAID(): Promise<string> {
+        return this.fetchMRAID().then(mraid => {
+            const markup = this._campaign.getDynamicMarkup();
+            if(markup) {
+                mraid = mraid.replace('{UNITY_DYNAMIC_MARKUP}', markup);
+            }
+
+            return MRAIDContainer.replace('<body></body>', '<body>' + mraid.replace('<script src="mraid.js"></script>', '') + '</body>');
+        });
+    }
+
     private updateProgressCircle(container: HTMLElement, value: number) {
         const wrapperElement = <HTMLElement>container.querySelector('.progress-wrapper');
 
@@ -269,12 +280,6 @@ export class MRAID extends View {
         }
     }
 
-    private createMRAID(): Promise<string> {
-        return this.fetchMRAID().then(mraid => {
-            return MRAIDContainer.replace('<body></body>', '<body>' + mraid.replace('<script src="mraid.js"></script>', '') + '</body>');
-        });
-    }
-
     private fetchMRAID(): Promise<string> {
         const resourceUrl = this._campaign.getResourceUrl();
         if(resourceUrl) {
@@ -282,14 +287,7 @@ export class MRAID extends View {
             if(fileId) {
                 return this._nativeBridge.Cache.getFileContent(fileId, 'UTF-8');
             } else {
-                return new Promise((resolve, reject) => {
-                    const xhr = new XMLHttpRequest();
-                    xhr.addEventListener('load', () => {
-                        resolve(xhr.responseText);
-                    }, false);
-                    xhr.open('GET', decodeURIComponent(resourceUrl.getUrl()));
-                    xhr.send();
-                });
+                throw new Error('Missing fileId');
             }
         } else {
             return Promise.resolve(this._campaign.getResource());